+++ cmsdev/internal/test/bos/bos_healthz.go
@@ -39,7 +39,7 @@ const bosV2HealthzUri = bosV2BaseUri + "/healthz"
 
 const bosV1HealthzCLI = "healthz"
 const bosV2HealthzCLI = "healthz"
-const bosDefaultHealthzCLI = bosV2HealthzCLI
+const bosDefaultHealthzCLI = bosV1HealthzCLI
 
 func healthzTestsAPI(params *common.Params) (passed bool) {
 	passed = true