+++ cmsdev/internal/test/bos/bos_options.go
@@ -38,7 +38,6 @@ import (
 const bosV2OptionsUri = bosV2BaseUri + "/options"
 
 const bosV2OptionsCLI = "options"
-const bosDefaultOptionsCLI = bosV2OptionsCLI
 
 func optionsTestsAPI(params *common.Params) (passed bool) {
 	passed = true
@@ -59,11 +58,6 @@ func optionsTestsCLI() (passed bool) {
 
 	// "v2 options list"
 	if !basicCLIListVerifyStringMapTest("v2", bosV2OptionsCLI) {
-		passed = false
-	}
-
-	// "options list"
-	if !basicCLIListVerifyStringMapTest(bosDefaultOptionsCLI) {
 		passed = false
 	}